Allow signalfx_dashboard_group.import_qualifier to be set to empty
import_qualifier is sometimes set to this in the API in a dashboard group:

```json
"importQualifiers" : [ {
  "filters" : [ ],
  "metric" : ""
} ]
```

This results in not being able to resolve to a clean plan because removing the entry in tf doesn't remove the entry in the API. That's because the field is set to omitempty which results in the field not being sent when encoded to JSON. (In our API if you don't include a field in a PUT request the field isn't modified which makes this harder to deal with).

With this change the plan will at least be clean when the empty resource is included in tf:

```
import_qualifiers {
}
```

This can be removed by sending a manual API request to update the dashboard group by setting `importQualifiers: []`. However if you modify the dashboard group in the UI the empty importQualifiers entry will return.

Resolves #252