Skip invalid target in datalinks
When changes are made to an existing target, the Terraform plugin SDK seems
to be creating an extraneous target with all empty values. Since name is a
required field on target, skip when we encounter empty names. Ideally this
issue would be fixed at the Terraform SDK level - this is only a workaround.