Improve the way dashboard group tells what mirrored dashboards are
Dashboard group is telling if a dashboard is mirrored or not based on
whether there is any override field or not in the dashboard config object.

This makes wrong judgements often times and skips adding mirrored dashboard
that should have been added to the dashboard group.

Then, this misjudgement leads to the existing dashboard in the config not being
recognized and being added over and over whenever the config is applied.

This change also improves dashboard trackability by keeping track of config_id
of dashboard configs.