Prevent dashboard_group resource from adding empty dashboards
This change fixes an issue that dashboard_group getting empty dashboards when
it has non-mirrored dashboards associated with it.

The root cause of the issue was an inaccurate length of dashboard slice of the dashboard group,
and this change fixes the issue by making the slice have the proper number of items.