Feature: Adding support for provider level tags (#581)
* Adding support for unique sets

Adding support for unique lists so that tag values from the provider can
be merged with those values from the specific resource.

* Adding support for provider supplied tags

Adding tags at the provider level allows for these to be applied to all
resources the provider configuration is managing.

* Validating against provider vnext

In order to help validate the changes earlier, this will allow the
provider to check the settings without needing to run a full integration
test

* Adding tag support for detectors vnext

* Adding support for provider level tags within current version

* Adding support for tags to all valid resources

This allows for resources to be configured the tag set

* Fixed test issue

* Simplifying OrderedSet with primitive slice